Register and test gym env
