+++ gym/cpr_gym_rs/cpr_gym_rs/__init__.py
@@ -1,0 +1,8 @@
+from gymnasium.envs.registration import register
+
+register(
+    id="cpr_gym_rs/FC16SSZwPT-v0",
+    entry_point="cpr_gym_rs.envs:FC16SSZwPT",
+    nondeterministic=True,
+    order_enforce=False,
+)
