+++ gym/cpr_gym_rs/pyproject.toml
@@ -14,3 +14,4 @@ dynamic = ["version"]
 
 [tool.maturin]
 features = ["pyo3/extension-module"]
+module-name = "cpr_gym_rs._rust"
